gh-223: sanitize all Azure API errors
